[fix] 修复BuildAssetBundles时，如果AOT补充元数据dll文件不存在则抛异常的问题。改成错误提示。

+++ Assets/Editor/HuaTuo/HuatuoEditorHelper.cs
@@ -150,6 +150,11 @@ namespace Huatuo
             foreach (var dll in aotDlls)
             {
                 string dllPath = $"{aotDllDir}/{dll}";
+                if(!File.Exists(dllPath))
+                {
+                    Debug.LogError($"ab中添加AOT补充元数据dll:{dllPath} 时发生错误,文件不存在。需要构建一次主包后才能生成裁剪后的AOT dll");
+                    continue;
+                }
                 string dllBytesPath = $"{tempDir}/{dll}.bytes";
                 File.Copy(dllPath, dllBytesPath, true);
                 notSceneAssets.Add(dllBytesPath);
